% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_pal.R
\docType{data}
\name{colors_delabj}
\alias{colors_delabj}
\title{Set of colors for delabj}
\format{An object of class \code{character} of length 15.}
\usage{
colors_delabj[name]
}
\value{
Named characters of the hex code
}
\description{
Simply a set of hex codes and names for colors I like
}
\keyword{datasets}
