% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_pal.R
\name{get_delabj_colors}
\alias{get_delabj_colors}
\title{Retrieve colors from delabj colors}
\usage{
get_delabj_colors(color1, color2)
}
\value{
a set of named characters
}
\description{
A function to retrieve a list of colors from the list of named colors in delabj_colors
}
\examples{
get_delabj_colors('plum', 'mint', 'zune pink')
}
