% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_pal.R
\name{get_delabj_palettes}
\alias{get_delabj_palettes}
\title{Retrieve Delabj color palettes}
\usage{
get_delabj_palettes(palette = "main", reverse = FALSE, ...)
}
\description{
Easily retrieve various palettes from the delabj collection.
}
