% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tweeks.R
\name{gridlines_major}
\alias{gridlines_major}
\title{Easily adjust ggplot2 major gridlines}
\usage{
gridlines_major(
  size = 0.5,
  color = "#d6d6d6",
  linetype = "solid",
  lineend = NULL,
  colour = NULL
)
}
\description{
A simple function to easily adjust the ggplot gridlines.
 This acts as a wrapper for the theme elements and is a little more user friendly.

 @usage
 gridlines_major(size = 0.5, color="#D6D6D6", linetype = "solid", lineend = NULL, colour =NULL )

 @return
 A list with the ggplot theme changes. This can be added to a ggplot object.

 @examples
 df <- data.frame(x = factor(rep(letters[1:3], each = 10)), y = rnorm(30), color=(rep(c("A", "B"), each=5)))
 ggplot(df, aes(x = x, y = y, color=color)) +
 geom_point()+
 gridlines_major(size=4, color = "pink")
 @export
}
