% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tweeks.R
\name{gridlines_off}
\alias{gridlines_off}
\title{Easily disable ggplot2 gridlines}
\usage{
'gridlines_off()'
'gridlines_major_off()'
'gridlines_minor_off()'
}
\description{
Quick functions to enable faster and more readable theme adjustments to ggplot2 when removing gridlines
Functions to do this are as follows:
* 'gridlines_off' disables all gridlines for a given plot
* 'gridlines_major_off' disables all major gridlines for a plot
* 'gridlines_minor_of' disables all minor gridlines for a plot
}
\examples{
library(ggplot2)

# Just major gridlines off
ggplot(mtcars, aes(x=mpg,y=hp))+
  geom_point()+
  gridlines_major_off()

# Just minor gridlines off
ggplot(mtcars, aes(x=mpg,y=hp))+
 geom_point()+
 gridlines_minor_off()

## All gridlines off
ggplot(mtcars, aes(x=mpg,y=hp))+
 geom_point()+
 gridlines_off()
 @export
}
\author{
Joshua de la Bruere
}
