% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.R
\name{import_poppins}
\alias{import_poppins}
\title{Import Fonts for Use in Plots and Charts}
\usage{
import_poppins()
import_source_code_pro()
}
\description{
A function to load the Poppins font for ease of use.
}
