% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tweeks.R
\name{legend_none}
\alias{legend_none}
\title{Easily position the ggplot legend}
\usage{
'legend_off()'
'legend_top()'
'legend_bottom()'
'legend_left()'
'legend_right()'
}
\description{
Quick functions to enable faster and more readable theme adjustments to ggplot2 when positioning the legend
Functons provided are:
* 'legend_off' disables the legend for a plot
* 'legend_top' puts the legened in the top posistion
* 'legend_bottom' puts the legened in the bottom posistion
* 'legend_left' puts the legened in the left posistion
* 'legend_right' puts the legened in the top posistion
}
\examples{
library(ggplot2)

# Just major gridlines off
ggplot(mtcars, aes(x=mpg,y=hp, colour = wt))+
  geom_point()+
  legend_off()

# Just minor gridlines off
ggplot(mtcars, aes(x=mpg,y=hp, colour = wt))+
 geom_point()+
 legend_top()

## All gridlines off
ggplot(mtcars, aes(x=mpg,y=hp, colour = wt))+
 geom_point()+
 legend_bottom()

 # Legend position
}
\author{
Joshua de la Bruere
}
