% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%not_in\%}
\alias{\%not_in\%}
\alias{\%xor\%}
\title{Operators}
\usage{
x \%not_in\% y

x \%xor\% y
}
\description{
@description
 Use these operators to do logical comparisons that fit with the tidyverse style pipe aesthetic
 @param x a value
 @param y a value
 @examples
 # not in
 c(1,2,4) %!in% c(1,24,6,5)
}
\details{
# x or
 # indicates elementwise exclusive OR.
 TRUE %xor% TRUE
 # FALSE

 FALSE %xor% FALSE
 # FALSE

 FALSE %xor% TRUE
 # TRUE
 @rdname operators
Not In
}
