% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_pal.R
\docType{data}
\name{palettes_delabj}
\alias{palettes_delabj}
\title{Delabj Color Pallets}
\format{An object of class \code{list} of length 2.}
\usage{
palettes_delabj[pallet_name]
}
\description{
A collection of pallets I use
}
\keyword{datasets}
