% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_pal.R
\name{scale_color_delabj}
\alias{scale_color_delabj}
\title{Change color to one of delabj's palettes}
\usage{
get_delabj_palettes(palette="main", reverse=  F, ...)
}
\arguments{
\item{palette}{Name of the palette (main, zune, retro)}

\item{reverse}{Boolean: should the palette be reversed}
}
\description{
Custom color scale to use the delabj palettes.
}
