% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_and_pal.R
\name{scale_color_zune}
\alias{scale_color_zune}
\title{A dedicated continuous palette for the zune colors}
\usage{
scale_color_zune(...)
}
\description{
Dedicated continuous scale using the zune colors.
The Zune was a mp3 player by Microsoft, with a distinctive gradient at the time. I was always a big fan of it and it makes a fairly nice continous scale as well.
}
