% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_delabj}
\alias{theme_delabj}
\title{My Prefered ggplot2 themes}
\usage{
theme_delabj(
base_size =12,
base_family = "Poppins",
base_line_size = .5,
base_rect_size = .5)

theme_delabj_dark(
base_size = 12,
base_family = "Poppins",
base_line_size = base_size/22,
base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family (Poppins by default)}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
A custom theme for ggplot2 based on theme_minimal(). A light and dark theme are both available.
}
\details{
theme_delabj has an off white background and light gridlines.
theme_delabj_dark has a dark grey background and dark gridlines.
}
\examples{
## Not Run
library(ggplot2)
df <- data.frame(x = factor(rep(letters[1:5], each = 10)), y = rnorm(50), color=(rep(c("A", "B", "C", "B", "A"), each=10)))
plot <- ggplot(df, aes(x = x, y = y, color=color)) + geom_jitter()


plot + theme_delabj

plot + theme_delabj_dark
}
